Addressing Cliff's comments
